+++ cookbooks/aws-parallelcluster-platform/spec/unit/recipes/supervisord_config_spec.rb
@@ -57,6 +57,28 @@ describe 'aws-parallelcluster-platform::supervisord_config' do
         end
       end
 
+      context "when head node and cfn-hup disabled on fleet" do
+        cached(:chef_run) do
+          runner = runner(platform: platform, version: version) do |node|
+            node.override['cluster']['node_type'] = 'HeadNode'
+            node.override['cluster']['dcv_enabled'] = 'head_node'
+            node.override['cluster']['in_place_update_on_fleet_enabled'] = 'false'
+            allow_any_instance_of(Object).to receive(:dcv_installed?).and_return(true)
+          end
+          runner.converge(described_recipe)
+        end
+        cached(:node) { chef_run.node }
+
+        it 'has the correct content' do
+          is_expected.to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
+            .with_content("[program:cfn-hup]")
+            .with_content("[program:clustermgtd]")
+            .with_content("[program:clusterstatusmgtd]")
+            .with_content("[program:pcluster_dcv_authenticator]")
+            .with_content("--port 8444")
+        end
+      end
+
       context "when compute fleet" do
         cached(:chef_run) do
           runner = runner(platform: platform, version: version) do |node|
@@ -75,6 +97,25 @@ describe 'aws-parallelcluster-platform::supervisord_config' do
 
           is_expected.not_to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
             .with_content("[program:pcluster_dcv_authenticator]")
+        end
+      end
+
+      context "when compute fleet with cfn-hup disabled on fleet" do
+        cached(:chef_run) do
+          runner = runner(platform: platform, version: version) do |node|
+            node.override['cluster']['node_type'] = 'ComputeFleet'
+            node.override['cluster']['in_place_update_on_fleet_enabled'] = 'false'
+          end
+          runner.converge(described_recipe)
+        end
+        cached(:node) { chef_run.node }
+
+        it 'has the correct content' do
+          is_expected.to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
+            .with_content("[program:computemgtd]")
+
+          is_expected.not_to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
+            .with_content("[program:cfn-hup]")
         end
       end
       context "when login node and dcv configured" do
@@ -109,10 +150,30 @@ describe 'aws-parallelcluster-platform::supervisord_config' do
 
           it 'has the correct content' do
             is_expected.to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
+              .with_content("[program:cfn-hup]")
               .with_content("[program:loginmgtd]")
 
             is_expected.not_to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
               .with_content("[program:pcluster_dcv_authenticator]")
+          end
+        end
+
+        context "when login node with cfn-hup disabled on fleet" do
+          cached(:chef_run) do
+            runner = runner(platform: platform, version: version) do |node|
+              node.override['cluster']['node_type'] = 'LoginNode'
+              node.override['cluster']['in_place_update_on_fleet_enabled'] = 'false'
+            end
+            runner.converge(described_recipe)
+          end
+          cached(:node) { chef_run.node }
+
+          it 'has the correct content' do
+            is_expected.to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
+              .with_content("[program:loginmgtd]")
+
+            is_expected.not_to render_file('/etc/parallelcluster/parallelcluster_supervisord.conf')
+              .with_content("[program:cfn-hup]")
           end
         end
       end